requirements.txt
requests
mitmproxy
pycryptodome